+++ src/Domain/Model/Rule.php
@@ -21,6 +21,7 @@ class Rule
     private ?int $emailAccessMinEvaluation = null;
     private ?int $sendEmailMinEvaluation = 5;
     private bool $requestAutomaticallyTransformedIntoMeeting = false;
+    private bool $disableMeetingRequest = false;
 
     public function __construct(
         Event $event,
@@ -144,19 +145,26 @@ class Rule
         return $this->requestAutomaticallyTransformedIntoMeeting;
     }
 
+    public function isMeetingRequestDisabled(): bool
+    {
+        return $this->disableMeetingRequest;
+    }
+
     public function update(
         array $what,
         $priority,
         ?int $phoneAccessMinEvaluation,
         ?int $emailAccessMinEvaluation,
         ?int $sendEmailMinEvaluation,
-        bool $requestAutomaticallyTransformedIntoMeeting = false
-    ) {
+        bool $requestAutomaticallyTransformedIntoMeeting,
+        bool $disableMeetingRequest
+    ): void {
         $this->what     = $what;
         $this->priority = $priority;
         $this->phoneAccessMinEvaluation = $phoneAccessMinEvaluation;
         $this->emailAccessMinEvaluation = $emailAccessMinEvaluation;
         $this->sendEmailMinEvaluation = $sendEmailMinEvaluation;
         $this->requestAutomaticallyTransformedIntoMeeting = $requestAutomaticallyTransformedIntoMeeting;
+        $this->disableMeetingRequest = $disableMeetingRequest;
     }
 }