+++ tests/Application/Components/Rule/ParticipantInfoAccessRulesResolverTest.php
@@ -44,6 +44,7 @@ class ParticipantInfoAccessRulesResolverTest extends TestCase
         $this->assertTrue($noRule->isPhoneVisible(1));
         $this->assertTrue($noRule->isEmailVisible(1));
         $this->assertFalse($noRule->canSendFollowUpEmail(1));
+        $this->assertTrue($noRule->canRequestMeeting());
     }
 
     public function testSingleRule()
@@ -75,6 +76,7 @@ class ParticipantInfoAccessRulesResolverTest extends TestCase
         $rule->getSendEmailMinEvaluation()->shouldBeCalled()->willReturn(4);
         $rule->getSeer()->shouldBeCalled()->willReturn($seerType->reveal());
         $rule->getSeeable()->shouldBeCalled()->willReturn($seeableType->reveal());
+        $rule->isMeetingRequestDisabled()->willReturn(true);
 
         $this->ruleRepository->getByEvent($event->reveal())->shouldBeCalled()->willReturn([$rule->reveal()]);
 
@@ -89,5 +91,6 @@ class ParticipantInfoAccessRulesResolverTest extends TestCase
         $this->assertTrue($singleAccessRule->isPhoneVisible(4));
         $this->assertTrue($singleAccessRule->isEmailVisible(3));
         $this->assertTrue($singleAccessRule->canSendFollowUpEmail(5));
+        $this->assertFalse($singleAccessRule->canRequestMeeting());
     }
 }